'''CORE:''' Large patch: Symbol/signet skills and char heading calculation fixes.
Symbol skills big thanks to Nuradi for much of the code and Rnn for info and testing.
Added or reworked the following skills:
Volcano, Cyclone, Gehenna, RagingWaves?, Anti-Summoning-Field, Symbol of the Assassin
Symbol of the Sniper, Symbol of Energy, Symbol of Honor, Symbol of Resistance,
Symbol of Noise, Symbol of Defense, Battle Force, Spell Force 

Thanks for code: L2j Server

+++ L2JTeon/java/net/sf/l2j/gameserver/skills/DocumentBase.java
@@ -375,7 +375,7 @@ abstract class DocumentBase
 	{
 		Condition cond = null;
 		int[] ElementSeeds = new int[5];
-		int[] forces = new int[2];
+		byte[] forces = new byte[2];
 		NamedNodeMap attrs = n.getAttributes();
 		for (int i = 0; i < attrs.getLength(); i++)
 		{
@@ -457,11 +457,11 @@ abstract class DocumentBase
 			}
 			else if ("battle_force".equalsIgnoreCase(a.getNodeName()))
 			{
-				forces[0] = Integer.decode(getValue(a.getNodeValue(), null));
+				forces[0] = Byte.decode(getValue(a.getNodeValue(), null));
 			}
 			else if ("spell_force".equalsIgnoreCase(a.getNodeName()))
 			{
-				forces[1] = Integer.decode(getValue(a.getNodeValue(), null));
+				forces[1] = Byte.decode(getValue(a.getNodeValue(), null));
 			}
 		}
 		// Elemental seed condition processing
